consolidate roi_utils in segmentation pipeline
