+++ src/ophys_etl/modules/segmentation/merge/candidates.py
@@ -10,7 +10,7 @@ import multiprocessing.managers
 from ophys_etl.modules.segmentation.merge.utils import (
     _winnow_process_list)
 from ophys_etl.modules.decrosstalk.ophys_plane import OphysROI
-from ophys_etl.modules.segmentation.merge.roi_utils import (
+from ophys_etl.modules.segmentation.utils.roi_utils import (
     do_rois_abut)
