+++ tests/modules/decrosstalk/test_ophys_plane.py
@@ -3,7 +3,6 @@ import json
 import numpy as np
 from ophys_etl.modules.decrosstalk.ophys_plane import OphysROI
 from ophys_etl.modules.decrosstalk.ophys_plane import DecrosstalkingOphysPlane
-from ophys_etl.modules.decrosstalk.ophys_plane import intersection_over_union
 
 from .utils import get_data_dir
 
@@ -71,42 +70,3 @@ def test_roi_global_pixel_set():
                    valid_roi=True,
                    mask_matrix=mask)
     assert roi.global_pixel_set == set([(202, 104), (203, 106)])
-
-
-def test_intersection_over_union():
-
-    width = 7
-    height = 5
-    mask = np.ones((height, width), dtype=bool)
-    mask[:, 4:] = False
-    roi0 = OphysROI(roi_id=0,
-                    x0=100,
-                    y0=200,
-                    width=width,
-                    height=height,
-                    valid_roi=True,
-                    mask_matrix=mask)
-
-    width = 4
-    height = 9
-    mask = np.ones((height, width), dtype=bool)
-    mask[:, 0] = False
-    mask[2:, :] = False
-    roi1 = OphysROI(roi_id=0,
-                    x0=101,
-                    y0=201,
-                    width=width,
-                    height=height,
-                    valid_roi=True,
-                    mask_matrix=mask)
-
-    # expected_intersection = 4
-    # expected_union = 22
-
-    expected = 4.0/22.0
-
-    actual = intersection_over_union(roi0, roi1)
-    actual1 = intersection_over_union(roi1, roi0)
-    eps = 1.0e-20
-    np.testing.assert_allclose(actual, actual1, rtol=0.0, atol=eps)
-    np.testing.assert_allclose(actual, expected, rtol=0.0, atol=eps)
